Did byrd stuff

+++ src/CSCI446_Project4_ReinforcementLearning/agent.h
@@ -14,14 +14,14 @@
 class Agent {
 public:
     Agent();
-    virtual Point next_accel(const Point& pos, const Point& vel, const int reward, const bool terminate) = 0;
+    virtual Point next_accel(const Point& pos, const Point& vel, const double reward, const bool terminate) = 0;
     
 };
 
 class DumbAgent : public Agent{
 public:
     DumbAgent();
-    Point next_accel(const Point& pos, const Point& vel, const int reward, const bool terminate);
+    Point next_accel(const Point& pos, const Point& vel, const double reward, const bool terminate);
     
 };
